journaldb trait moved to the lib.rs file
