+++ util/journaldb/src/overlaydb.rs
@@ -135,10 +135,9 @@ impl OverlayDB {
 
 	/// Revert all operations on this object (i.e. `insert()`s and `remove()`s) since the
 	/// last `commit()`.
-	pub fn revert(&mut self) { self.overlay.clear(); }
-
-	/// Get the number of references that would be committed.
-	pub fn commit_refs(&self, key: &H256) -> i32 { self.overlay.raw(key, EMPTY_PREFIX).map_or(0, |(_, refs)| refs) }
+	pub fn revert(&mut self) {
+		self.overlay.clear();
+	}
 
 	/// Get the refs and value of the given key.
 	fn payload(&self, key: &H256) -> Option<Payload> {
